Add salesPrice and name to order items
Updated the order and item schemas and interfaces in the backend and frontend to store item names and sale prices. Enhanced item addition function to include the name and sale price of items when adding to an order. This will provide more information about items in an order and will be helpful in the calculation of total order cost.