Add function to update item in an order
A function called `updatedItemInOrder` was added to the order controller to enable updating the details of an item in an given order. An endpoint, `/:orderId/updateItem/:itemId`, was also created in the orders router to access this newly created function. Now, one can easily update the amount, sales price, VAT, and discount of an item in an order.